feat(idea-reworked): split file and value parsing
File parsing is not very much related to value parsing: values
are (mostly) interpreted by *TeX and not natbib/biber, apart
from " and " in list values. This is clearly demonstrated by
the following example of two entries producing the same output:

```
@book{a, title = "About the \copyright symbol"}
@book{b, title = "About the \copy" # "right symbol"}
```

While this is a gimmick, so to speak, it does demonstrate that
you cannot interpret or even tokenize TeX commands, including
Unicode characters and scripts without first parsing the file.
This distinguishes file and value syntax, which leads to
separate grammars (one for "BibTeX", one for basic TeX + list
values).

This also has the added benefit of having a separate value
parser for Bib.TXT, and fixes a bunch of other problems.

BREAKING CHANGE: basic parse does not parse values anymore