fix(idea-reworked): fix fallback for sub/sup
Fix the HTML fallback for sub- and superscript in math mode.

Most characters that are made sub- or superscript in math mode BibTeX 
are numbers, which have code points in Unicode (as well as some 
letters). However, if a formatted string does not only contain supported 
characters, it reverts the entire string to HTML instead, to keep 
formatting consistent.